fix: exported interface
