+++ index.js
@@ -19,3 +19,5 @@ function responseSchemaCompiler (fjsOpts, { schema /* method, url, httpStatus */
 
 module.exports = SerializerSelector
 module.exports.default = SerializerSelector
+module.exports.SerializerSelector = SerializerSelector
+module.exports.StandaloneSerializer = require('./standalone')
