+++ test/types/index.test-d.ts
@@ -1,6 +1,16 @@
 import { expectType } from "tsd";
-import SerializerSelector, { SerializerCompiler } from "../..";
+import SerializerSelector, {
+  SerializerCompiler,
+  SerializerSelector as SerializerSelectorNamed,
+  StandaloneSerializer,
+} from "../..";
 
-const compiler = SerializerSelector();
+{
+  const compiler = SerializerSelector();
+  expectType<SerializerCompiler>(compiler);
+}
 
-expectType<SerializerCompiler>(compiler);
+{
+  const compiler = SerializerSelectorNamed();
+  expectType<SerializerCompiler>(compiler);
+}
